vick part registeration finished
