+++ src/AppRoutes.tsx
@@ -33,9 +33,15 @@ const CarrierRegister = lazy(() => import("./pages/register/carrier"));
 const CarrierRegisterList = lazy(
   () => import("./pages/register/carrier/CarrierRegisterList")
 );
-const CarrierRegisterForm = lazy(() => import("./pages/register/carrier/CarrierRegisterForm"))
-const CarrierProfile = lazy(() => import("./pages/register/carrier/CarrierProfile"))
-const CarrierContactInformation = lazy(() => import("./pages/register/carrier/CarrierContactInformation"))
+const CarrierRegisterForm = lazy(
+  () => import("./pages/register/carrier/CarrierRegisterForm")
+);
+const CarrierProfile = lazy(
+  () => import("./pages/register/carrier/CarrierProfile")
+);
+const CarrierContactInformation = lazy(
+  () => import("./pages/register/carrier/CarrierContactInformation")
+);
 
 // vendor
 const VendorRegister = lazy(() => import("./pages/register/vendor"));
@@ -49,10 +55,10 @@ const VendorDetails = lazy(
   () => import("./pages/register/vendor/vendorDetails")
 );
 const VendorProfile = lazy(
-  () => import("./pages/register/vendor/vendorDetails/vendorProfile")
+  () => import("./pages/register/vendor/vendorDetails/VendorProfile")
 );
 const VendorContact = lazy(
-  () => import("./pages/register/vendor/vendorDetails/vendorContact")
+  () => import("./pages/register/vendor/vendorDetails/VendorContact")
 );
 
 // HRM Routes
@@ -120,7 +126,10 @@ const AppRoutes: React.FC = () => {
           <Route index element={<CarrierRegisterList />} />
           <Route path="register-form" element={<CarrierRegisterForm />} />
           <Route path="carrier-details/profile" element={<CarrierProfile />} />
-          <Route path="carrier-details/contact-information" element={<CarrierContactInformation />} />
+          <Route
+            path="carrier-details/contact-information"
+            element={<CarrierContactInformation />}
+          />
         </Route>
         <Route path="registration-vendor" element={<VendorRegister />}>
           <Route index element={<VendorRegisterList />} />
